Map_Menu -> MapMenu (trust me it makes sense)
The CamelCase translates to the snake_case filename. In general, it
makes little sense to combine both camel casing and underscores. In this
repository, the underscore is being used to group related labels
together, much like how you'd see "SomeModule.Something" in object
oriented code. So, unless there's a "Map" group that makes sense, I'm
fine with calling this group "MapMenu", so "Map" can be reserved for the
actual maps.